fix(build): 清零 Kotlin 编译告警
- PronunciationPlayer：TTS 旧版 onError 回调补 @Suppress("OVERRIDE_DEPRECATION")，基类抽象方法仍需实现
- StudyScreen：空态词典图标改用 Icons.AutoMirrored.Filled.MenuBook，从右到左布局下自动镜像
- TranslateScreen：弃用的 LocalClipboardManager 改为系统 ClipboardManager + ClipData
- AboutOverlayHost：LocalLifecycleOwner 改从 androidx.lifecycle.compose 取；强制浮层的返回键拦截去掉未使用表达式
- DictionaryApp：去掉 Int 上多余的 toInt() 转换
- 同步更新 WhatsNewData.kt

+++ app/src/main/java/com/chloemlla/cdict/ui/DictionaryApp.kt
@@ -579,7 +579,7 @@ private fun WordList(
                             itemsIndexed(state.words, key = { _, word -> word.id }) { index, word ->
                                 // Stagger: each item fades in with a tiny vertical slide, capped so
                                 // long lists don't drag the entrance past the animation duration.
-                                val itemDelay = (index.coerceAtMost(8) * 24).toInt()
+                                val itemDelay = index.coerceAtMost(8) * 24
                                 val itemAlpha by animateFloatAsState(
                                     targetValue = if (entranceProgress.value >= 1f ||
                                         entranceProgress.value >= (itemDelay + 1) / 281f
